card_news_generator/requirements.txt
pillow>=12.0.0
jinja2>=3.1.0
anthropic>=0.39.0
# playwright>=1.40.0  # 선택: HTML 기반 고품질 렌더링 시 필요
